Move tests for insert/update operations to test_ingester
